05_prediction/requirements.txt
flask
joblib
pandas
numpy
xgboost
scikit-learn
psycopg2-binary
python-dotenv